@@ -14,7 +14,7 @@ description: try to wirte all widget
 version: 1.0.0+1
 
 environment:
-  sdk: ">=2.1.0 <3.0.0"
+  sdk: ">=2.16.2 <3.0.0"
 
 dependencies:
   flutter:
@@ -23,13 +23,12 @@ dependencies:
   # The following adds the Cupertino Icons font to your application.
   # Use with the CupertinoIcons class for iOS style icons.
   cupertino_icons: ^0.1.2
-  english_words: ^3.1.0
-  photo_view: ^0.6.0
-  cached_network_image: ^0.7.0
-  keyboard_visibility: ^0.5.6
-  provider: ^3.0.0+1
-  charts_flutter: ^0.9.0
-  fl_chart: ^0.7.0
+  english_words: ^4.0.0
+  photo_view: ^0.14.0
+  cached_network_image: ^3.2.1
+  charts_flutter: ^0.12.0
+  fl_chart: ^0.55.0
+  provider: ^6.0.3
 dev_dependencies:
   flutter_test:
     sdk: flutter
