@@ -7,9 +7,9 @@ class AnimatedListViewWidget extends StatefulWidget {
 
 class _AnimatedListViewWidgetState extends State<AnimatedListViewWidget> {
   final GlobalKey<AnimatedListState> _listKey = GlobalKey<AnimatedListState>();
-  ListModel<int> _list;
-  int _selectedItem;
-  int _nextItem; // The next item inserted when the user presses the '+' button.
+  late ListModel<int> _list;
+  int _selectedItem = -1;
+  int _nextItem = 0; // The next item inserted when the user presses the '+' button.
 
   @override
   void initState() {
@@ -31,7 +31,7 @@ class _AnimatedListViewWidgetState extends State<AnimatedListViewWidget> {
       selected: _selectedItem == _list[index],
       onTap: () {
         setState(() {
-          _selectedItem = _selectedItem == _list[index] ? null : _list[index];
+          _selectedItem = (_selectedItem == _list[index] ? null : _list[index])!;
         });
       },
     );
@@ -47,7 +47,7 @@ class _AnimatedListViewWidgetState extends State<AnimatedListViewWidget> {
     return CardItem(
       animation: animation,
       item: item,
-      selected: false,
+      selected: false, onTap: () {  },
       // No gesture detector here: we don't want removed items to be interactive.
     );
   }
@@ -64,7 +64,7 @@ class _AnimatedListViewWidgetState extends State<AnimatedListViewWidget> {
     if (_selectedItem != null) {
       _list.removeAt(_list.indexOf(_selectedItem));
       setState(() {
-        _selectedItem = null;
+        _selectedItem = -1;
       });
     }
   }
@@ -103,9 +103,9 @@ class _AnimatedListViewWidgetState extends State<AnimatedListViewWidget> {
 }
 class ListModel<E> {
   ListModel({
-    @required this.listKey,
-    @required this.removedItemBuilder,
-    Iterable<E> initialItems,
+    required this.listKey,
+    required this.removedItemBuilder,
+    required Iterable<E> initialItems,
   })  : assert(listKey != null),
         assert(removedItemBuilder != null),
         _items = List<E>.from(initialItems ?? <E>[]);
@@ -114,17 +114,17 @@ class ListModel<E> {
   final dynamic removedItemBuilder;
   final List<E> _items;
 
-  AnimatedListState get _animatedList => listKey.currentState;
+  AnimatedListState? get _animatedList => listKey.currentState;
 
   void insert(int index, E item) {
     _items.insert(index, item);
-    _animatedList.insertItem(index);
+    _animatedList?.insertItem(index);
   }
 
   E removeAt(int index) {
     final E removedItem = _items.removeAt(index);
     if (removedItem != null) {
-      _animatedList.removeItem(index,
+      _animatedList?.removeItem(index,
               (BuildContext context, Animation<double> animation) {
             return removedItemBuilder(removedItem, context, animation);
           });
@@ -143,10 +143,10 @@ class ListModel<E> {
 
 class CardItem extends StatelessWidget {
   const CardItem(
-      {Key key,
-        @required this.animation,
-        this.onTap,
-        @required this.item,
+      { Key? key,
+        required this.animation,
+        required this.onTap,
+        required this.item,
         this.selected: false})
       : assert(animation != null),
         assert(item != null && item >= 0),
@@ -160,9 +160,9 @@ class CardItem extends StatelessWidget {
 
   @override
   Widget build(BuildContext context) {
-    TextStyle textStyle = Theme.of(context).textTheme.display1;
+    TextStyle? textStyle = Theme.of(context).textTheme.displayMedium;
     if (selected)
-      textStyle = textStyle.copyWith(color: Colors.lightGreenAccent[400]);
+      textStyle = textStyle?.copyWith(color: Colors.lightGreenAccent[400]);
     return Padding(
       padding: const EdgeInsets.all(2.0),
       child: SizeTransition(
