@@ -27,7 +27,7 @@ class _ParentWidgetState extends State<ParentWidget> {
 //------------------------- TapboxB ----------------------------------
 
 class TapboxB extends StatelessWidget {
-  TapboxB({Key key, this.active: false, @required this.onChanged})
+  TapboxB({Key? key, this.active: false, required this.onChanged})
       : super(key: key);
 
   final bool active;
