@@ -2,7 +2,7 @@ import 'package:flutter/material.dart';
 
 class MyNotification extends Notification{
 
-  MyNotification({@required this.msg});
+  MyNotification({required this.msg});
 
   final String msg;
 }
