@@ -21,108 +21,109 @@ TextStyle secondStyle = TextStyle(fontSize: 16);
 TextStyle thirdStyle = TextStyle(fontSize: 12);
 
 class Entry {
-  Entry(this.title, [this.textStyle, this.children = const <Entry>[]]);
 
+  Entry({required this.title, required this.textStyle,  this.children=const []});
   final String title;
-  final List<Entry> children;
-  TextStyle textStyle = thirdStyle;
+  final TextStyle textStyle;
+
+  List<Entry> children = [];
+
+
 }
 
 // The entire multilevel list displayed by this app.
 final List<Entry> data = <Entry>[
   Entry(
-    'Element',
-    firstStyle,
-    <Entry>[
+    title: 'Element',
+    textStyle:  firstStyle,
+    children: <Entry>[
       Entry(
-        'Form',
-        secondStyle,
-        <Entry>[
-          Entry('Checkbox', thirdStyle),
-          Entry('Button', thirdStyle),
-          Entry('Slider', thirdStyle),
-          Entry('Switch', thirdStyle),
+       title: 'Form',
+        textStyle: secondStyle,
+        children: <Entry>[
+          Entry(title:'Checkbox',textStyle:  thirdStyle),
+          Entry(title:'Button',textStyle:  thirdStyle),
+          Entry(title:'Slider',textStyle:  thirdStyle),
+          Entry(title:'Switch',textStyle:  thirdStyle),
         ],
       ),
-      Entry('Frame', secondStyle, <Entry>[
-        Entry('Row', thirdStyle),
-        Entry('Column', thirdStyle),
-        Entry('Stack', thirdStyle),
-        Entry('Layout', thirdStyle),
-        Entry('Box', thirdStyle),
-        Entry('Expanded', thirdStyle),
-        Entry('Spacing', thirdStyle),
-        Entry('Table', thirdStyle),
+      Entry(title:'Frame', textStyle:secondStyle,children: <Entry>[
+        Entry(title:'Row',textStyle: thirdStyle),
+        Entry(title:'Column',textStyle: thirdStyle),
+        Entry(title:'Stack', textStyle:thirdStyle),
+        Entry(title:'Layout', textStyle:thirdStyle),
+        Entry(title:'Box', textStyle:thirdStyle),
+        Entry(title:'Expanded',textStyle: thirdStyle),
+        Entry(title:'Spacing', textStyle:thirdStyle),
+        Entry(title:'Table', textStyle:thirdStyle),
       ]),
-      Entry('Media', secondStyle, <Entry>[
-        Entry('Canvas', thirdStyle),
+      Entry(title:'Media',textStyle: secondStyle, children:<Entry>[
+        Entry(title:'Canvas',textStyle: thirdStyle),
       ]),
     ],
   ),
   Entry(
-    'Components',
-    firstStyle,
-    <Entry>[
-
-      Entry('List', secondStyle, <Entry>[
-        Entry('ListView', thirdStyle),
-        Entry('ListBody', thirdStyle),
-        Entry('AnimatedList', thirdStyle),
+    title: 'Components',
+   textStyle: firstStyle,
+  children:  <Entry>[
+      Entry(title:'List', textStyle:secondStyle,children: <Entry>[
+        Entry(title:'ListView', textStyle:thirdStyle),
+        Entry(title:'ListBody', textStyle:thirdStyle),
+        Entry(title:'AnimatedList', textStyle:thirdStyle),
       ]),
-      Entry('Bar', secondStyle, <Entry>[
-        Entry('AppBar', thirdStyle),
-        Entry('SnackBar', thirdStyle),
-        Entry('SliverAppBar', thirdStyle),
-        Entry('FlexbleSpaceBar', thirdStyle),
-        Entry('SnackBarAction', thirdStyle),
+      Entry(title:'Bar',textStyle: secondStyle,children :<Entry>[
+        Entry(title:'AppBar', textStyle:thirdStyle),
+        Entry(title:'SnackBar',textStyle: thirdStyle),
+        Entry(title:'SliverAppBar',textStyle: thirdStyle),
+        Entry(title:'FlexbleSpaceBar',textStyle: thirdStyle),
+        Entry(title:'SnackBarAction', textStyle:thirdStyle),
       ]),
-      Entry('Dialog', secondStyle, <Entry>[
-        Entry('AlertDialog', thirdStyle),
-        Entry('Dialog', thirdStyle),
-        Entry('AboutDialog', thirdStyle),
-        Entry('SimpleDialog', thirdStyle),
+      Entry(title:'Dialog', textStyle:secondStyle,children: <Entry>[
+        Entry(title:'AlertDialog',textStyle: thirdStyle),
+        Entry(title:'Dialog',textStyle: thirdStyle),
+        Entry(title:'AboutDialog',textStyle: thirdStyle),
+        Entry(title:'SimpleDialog',textStyle: thirdStyle),
       ]),
-      Entry('Grid', secondStyle, <Entry>[
-        Entry('GridTile', thirdStyle),
-        Entry('GridView', thirdStyle),
-        Entry('GridPaper', thirdStyle),
-        Entry('SliverGrid', thirdStyle),
-        Entry('GridTileBar', thirdStyle),
+      Entry(title:'Grid',textStyle: secondStyle,children: <Entry>[
+        Entry(title:'GridTile', textStyle:thirdStyle),
+        Entry(title:'GridView',textStyle: thirdStyle),
+        Entry(title:'GridPaper',textStyle: thirdStyle),
+        Entry(title:'SliverGrid',textStyle: thirdStyle),
+        Entry(title:'GridTileBar',textStyle: thirdStyle),
       ]),
-      Entry('Scroll', secondStyle, <Entry>[
-        Entry('ScrollMetrics', thirdStyle),
-        Entry('ScrollPhysics', thirdStyle),
+      Entry(title:'Scroll', textStyle:secondStyle,children: <Entry>[
+        Entry(title:'ScrollMetrics',textStyle: thirdStyle),
+        Entry(title:'ScrollPhysics',textStyle: thirdStyle),
       ]),
-      Entry('Menu', secondStyle, <Entry>[
-        Entry('CheckedPopupMenuItem', thirdStyle),
-        Entry('DropdownMenuItem', thirdStyle),
-        Entry('PopupMenuButton', thirdStyle),
-        Entry('PopupMenuDivider', thirdStyle),
-        Entry('PopupMenuEntry', thirdStyle),
-        Entry('PopupMenuItem', thirdStyle),
+      Entry(title:'Menu', textStyle:secondStyle, children:<Entry>[
+        Entry(title:'CheckedPopupMenuItem', textStyle:thirdStyle),
+        Entry(title:'DropdownMenuItem',textStyle: thirdStyle),
+        Entry(title:'PopupMenuButton',textStyle: thirdStyle),
+        Entry(title:'PopupMenuDivider',textStyle: thirdStyle),
+        Entry(title:'PopupMenuEntry',textStyle: thirdStyle),
+        Entry(title:'PopupMenuItem', textStyle:thirdStyle),
       ]),
-      Entry('Pick', secondStyle, <Entry>[
-        Entry('DayPicker', thirdStyle),
+      Entry(title:'Pick',textStyle: secondStyle,children: <Entry>[
+        Entry(title:'DayPicker',textStyle: thirdStyle),
       ]),
-      Entry('Chip', secondStyle, <Entry>[
-        Entry('Chip', thirdStyle),
-        Entry('ChipTheme', thirdStyle),
-        Entry('ChipThemeData', thirdStyle),
-        Entry('ChoiceChip', thirdStyle),
-        Entry('FilterChip', thirdStyle),
-        Entry('InputChip', thirdStyle),
-        Entry('RawChip', thirdStyle),
+      Entry(title:'Chip',textStyle: secondStyle,children: <Entry>[
+        Entry(title:'Chip', textStyle:thirdStyle),
+        Entry(title:'ChipTheme',textStyle: thirdStyle),
+        Entry(title:'ChipThemeData',textStyle: thirdStyle),
+        Entry(title:'ChoiceChip',textStyle: thirdStyle),
+        Entry(title:'FilterChip',textStyle: thirdStyle),
+        Entry(title:'InputChip',textStyle: thirdStyle),
+        Entry(title:'RawChip',textStyle: thirdStyle),
       ]),
-      Entry('Progress', secondStyle, <Entry>[
-        Entry('LinearProgressIndicator', thirdStyle),
-        Entry('CircularProgressIndicator', thirdStyle),
+      Entry(title:'Progress', textStyle:secondStyle,children: <Entry>[
+        Entry(title:'LinearProgressIndicator',textStyle: thirdStyle),
+        Entry(title:'CircularProgressIndicator',textStyle: thirdStyle),
       ]),
     ],
   ),
-
-  Entry("Other",firstStyle,<Entry>[
-    Entry("FutureBuild",thirdStyle),
-    Entry("Animation",thirdStyle),
+  Entry(title:"Other",textStyle: firstStyle,children: <Entry>[
+    Entry(title:"FutureBuild",textStyle: thirdStyle),
+    Entry(title:"Animation",textStyle: thirdStyle),
   ]),
 ];
 
@@ -151,7 +152,7 @@ class EntryItem extends StatelessWidget {
             style: root.textStyle,
           ),
         ),
-        children: root.children.map((it)=>_buildTiles(it,context)).toList(),
+        children: root.children.map((it) => _buildTiles(it, context)).toList(),
       );
     }
   }
