@@ -1,11 +1,11 @@
 class User<T>{
-   String name;
-   String email;
-   List<T> rate;
+   String name="";
+   String email="";
+   List<T> rate=[];
 
 }
 
 class Rate{
-  int time;
-  int amount;
+  int time =0;
+  int amount=0;
 }
